move managedFinalizeContract to negotiate.go
